linkedin_chatbot/requirements.txt
# Generador de Posts de LinkedIn con OpenAI

openai>=1.12.0
pydantic>=2.0.0
python-dotenv>=1.0.0
